+++ src/main/kotlin/baseball/Application.kt
@@ -1,5 +1,7 @@
 package baseball
 
+import baseball.controller.BaseballActivity
+
 fun main() {
     BaseballActivity()
 }
