+++ src/main/kotlin/baseball/controller/BaseballActivity.kt
@@ -1,7 +1,9 @@
-package baseball
+package baseball.controller
+
+import baseball.view.OutputView
 
 class BaseballActivity {
-    private val outputView:OutputView = OutputView()
+    private val outputView: OutputView = OutputView()
     init {
         outputView.gameStart()
         outputView.baseballGame()
