+++ src/main/kotlin/baseball/model/Judgment.kt
@@ -1,4 +1,4 @@
-package baseball
+package baseball.model
 
 class Judgment {
