+++ src/main/kotlin/baseball/model/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package baseball
+package baseball.model
 
 import camp.nextstep.edu.missionutils.Randoms
