+++ src/test/kotlin/baseball/InputValidateTest.kt
@@ -1,7 +1,7 @@
 package baseball
 
+import baseball.view.InputValidate
 import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
