+++ src/test/kotlin/baseball/JudgmentTest.kt
@@ -1,8 +1,8 @@
 package baseball
 
+import baseball.model.Judgment
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
