+++ src/test/kotlin/baseball/OutputViewTest.kt
@@ -1,5 +1,6 @@
 package baseball
 
+import baseball.view.OutputView
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.params.ParameterizedTest
